pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-verilog_env"
version = "0.1.0"
description = "EDA Hardware Optimization environment for OpenEnv — RTL-to-GDSII POMDP"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "verilog_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["verilog_env", "verilog_env.server", "verilog_env.server.tool_wrappers", "verilog_env.graders"]

[tool.setuptools.package-dir]
"verilog_env" = "."
"verilog_env.server" = "server"
"verilog_env.server.tool_wrappers" = "server/tool_wrappers"
"verilog_env.graders" = "graders"